refactor: Input들 각 단위로 분리
